Player Movement and ShadingOverlay deletion tiles
When the player clicks once, the player enters a "player_select_movement_state",  in this state, the player can view all possible moves. The player then clicks again to move to a VIABLE tile. If the tile isnt viable, it makes the player select again. After selecting a tile, the player will teleport to the tile and then remove all flood pathfinding overlay tiles.